[Core] Move To Util Package
