+++ core/src/main/java/com/example/core/util/ApiConst.kt
@@ -1,4 +1,4 @@
-package com.example.core
+package com.example.core.util
 
 const val SUCCESS = 200
 const val BASE_URL = "http://43.203.50.18:8000"
